chore: refacto by separating base model
