+++ pytorch_tabnet/abstract_model.py
@@ -1,0 +1,494 @@
+import torch
+import numpy as np
+from scipy.sparse import csc_matrix
+import time
+from abc import abstractmethod
+from pytorch_tabnet import tab_network
+from pytorch_tabnet.multiclass_utils import unique_labels
+from sklearn.metrics import roc_auc_score, mean_squared_error, accuracy_score
+from torch.nn.utils import clip_grad_norm_
+from pytorch_tabnet.utils import (PredictDataset,
+                                  create_dataloaders,
+                                  create_explain_matrix)
+from sklearn.base import BaseEstimator
+from torch.utils.data import DataLoader
+from copy import deepcopy
+import io
+import json
+from pathlib import Path
+import shutil
+import zipfile
+
+
+class TabModel(BaseEstimator):
+    def __init__(self, n_d=8, n_a=8, n_steps=3, gamma=1.3, cat_idxs=[], cat_dims=[], cat_emb_dim=1,
+                 n_independent=2, n_shared=2, epsilon=1e-15,  momentum=0.02,
+                 lambda_sparse=1e-3, seed=0,
+                 clip_value=1, verbose=1,
+                 optimizer_fn=torch.optim.Adam,
+                 optimizer_params=dict(lr=2e-2),
+                 scheduler_params=None, scheduler_fn=None,
+                 mask_type="sparsemax",
+                 input_dim=None, output_dim=None,
+                 device_name='auto'):
+        """ Class for TabNet model
+
+        Parameters
+        ----------
+            device_name: str
+                'cuda' if running on GPU, 'cpu' if not, 'auto' to autodetect
+        """
+
+        self.n_d = n_d
+        self.n_a = n_a
+        self.n_steps = n_steps
+        self.gamma = gamma
+        self.cat_idxs = cat_idxs
+        self.cat_dims = cat_dims
+        self.cat_emb_dim = cat_emb_dim
+        self.n_independent = n_independent
+        self.n_shared = n_shared
+        self.epsilon = epsilon
+        self.momentum = momentum
+        self.lambda_sparse = lambda_sparse
+        self.clip_value = clip_value
+        self.verbose = verbose
+        self.optimizer_fn = optimizer_fn
+        self.optimizer_params = optimizer_params
+        self.device_name = device_name
+        self.scheduler_params = scheduler_params
+        self.scheduler_fn = scheduler_fn
+        self.mask_type = mask_type
+        self.input_dim = input_dim
+        self.output_dim = output_dim
+
+        self.batch_size = 1024
+
+        self.seed = seed
+        torch.manual_seed(self.seed)
+        # Defining device
+        if device_name == 'auto':
+            if torch.cuda.is_available():
+                device_name = 'cuda'
+            else:
+                device_name = 'cpu'
+        self.device = torch.device(device_name)
+        print(f"Device used : {self.device}")
+
+    @abstractmethod
+    def construct_loaders(self, X_train, y_train, X_valid, y_valid,
+                          weights, batch_size, num_workers, drop_last):
+        """
+        Returns
+        -------
+        train_dataloader, valid_dataloader : torch.DataLoader, torch.DataLoader
+            Training and validation dataloaders
+        -------
+        """
+        raise NotImplementedError('users must define construct_loaders to use this base class')
+
+    def init_network(
+                     self,
+                     input_dim,
+                     output_dim,
+                     n_d,
+                     n_a,
+                     n_steps,
+                     gamma,
+                     cat_idxs,
+                     cat_dims,
+                     cat_emb_dim,
+                     n_independent,
+                     n_shared,
+                     epsilon,
+                     virtual_batch_size,
+                     momentum,
+                     device_name,
+                     mask_type,
+                     ):
+        self.network = tab_network.TabNet(
+            input_dim,
+            output_dim,
+            n_d=n_d,
+            n_a=n_a,
+            n_steps=n_steps,
+            gamma=gamma,
+            cat_idxs=cat_idxs,
+            cat_dims=cat_dims,
+            cat_emb_dim=cat_emb_dim,
+            n_independent=n_independent,
+            n_shared=n_shared,
+            epsilon=epsilon,
+            virtual_batch_size=virtual_batch_size,
+            momentum=momentum,
+            device_name=device_name,
+            mask_type=mask_type).to(self.device)
+
+        self.reducing_matrix = create_explain_matrix(
+            self.network.input_dim,
+            self.network.cat_emb_dim,
+            self.network.cat_idxs,
+            self.network.post_embed_dim)
+
+    def fit(self, X_train, y_train, X_valid=None, y_valid=None, loss_fn=None,
+            weights=0, max_epochs=100, patience=10, batch_size=1024,
+            virtual_batch_size=128, num_workers=0, drop_last=False):
+        """Train a neural network stored in self.network
+        Using train_dataloader for training data and
+        valid_dataloader for validation.
+
+        Parameters
+        ----------
+            X_train: np.ndarray
+                Train set
+            y_train : np.array
+                Train targets
+            X_train: np.ndarray
+                Train set
+            y_train : np.array
+                Train targets
+            weights : bool or dictionnary
+                0 for no balancing
+                1 for automated balancing
+                dict for custom weights per class
+            max_epochs : int
+                Maximum number of epochs during training
+            patience : int
+                Number of consecutive non improving epoch before early stopping
+            batch_size : int
+                Training batch size
+            virtual_batch_size : int
+                Batch size for Ghost Batch Normalization (virtual_batch_size < batch_size)
+            num_workers : int
+                Number of workers used in torch.utils.data.DataLoader
+            drop_last : bool
+                Whether to drop last batch during training
+        """
+        # update model name
+
+        self.update_fit_params(X_train, y_train, X_valid, y_valid, loss_fn,
+                               weights, max_epochs, patience, batch_size,
+                               virtual_batch_size, num_workers, drop_last)
+
+        train_dataloader, valid_dataloader = self.construct_loaders(X_train,
+                                                                    y_train,
+                                                                    X_valid,
+                                                                    y_valid,
+                                                                    self.updated_weights,
+                                                                    self.batch_size,
+                                                                    self.num_workers,
+                                                                    self.drop_last)
+
+        self.init_network(
+            input_dim=self.input_dim,
+            output_dim=self.output_dim,
+            n_d=self.n_d,
+            n_a=self.n_a,
+            n_steps=self.n_steps,
+            gamma=self.gamma,
+            cat_idxs=self.cat_idxs,
+            cat_dims=self.cat_dims,
+            cat_emb_dim=self.cat_emb_dim,
+            n_independent=self.n_independent,
+            n_shared=self.n_shared,
+            epsilon=self.epsilon,
+            virtual_batch_size=self.virtual_batch_size,
+            momentum=self.momentum,
+            device_name=self.device_name,
+            mask_type=self.mask_type
+        )
+
+        self.optimizer = self.optimizer_fn(self.network.parameters(),
+                                           **self.optimizer_params)
+
+        if self.scheduler_fn:
+            self.scheduler = self.scheduler_fn(self.optimizer, **self.scheduler_params)
+        else:
+            self.scheduler = None
+
+        self.losses_train = []
+        self.losses_valid = []
+        self.learning_rates = []
+        self.metrics_train = []
+        self.metrics_valid = []
+
+        if self.verbose > 0:
+            print("Will train until validation stopping metric",
+                  f"hasn't improved in {self.patience} rounds.")
+            msg_epoch = f'| EPOCH |  train  |   valid  | total time (s)'
+            print('---------------------------------------')
+            print(msg_epoch)
+
+        total_time = 0
+        while (self.epoch < self.max_epochs and
+               self.patience_counter < self.patience):
+            starting_time = time.time()
+            # updates learning rate history
+            self.learning_rates.append(self.optimizer.param_groups[-1]["lr"])
+
+            fit_metrics = self.fit_epoch(train_dataloader, valid_dataloader)
+
+            # leaving it here, may be used for callbacks later
+            self.losses_train.append(fit_metrics['train']['loss_avg'])
+            self.losses_valid.append(fit_metrics['valid']['total_loss'])
+            self.metrics_train.append(fit_metrics['train']['stopping_loss'])
+            self.metrics_valid.append(fit_metrics['valid']['stopping_loss'])
+
+            stopping_loss = fit_metrics['valid']['stopping_loss']
+            if stopping_loss < self.best_cost:
+                self.best_cost = stopping_loss
+                self.patience_counter = 0
+                # Saving model
+                self.best_network = deepcopy(self.network)
+            else:
+                self.patience_counter += 1
+
+            self.epoch += 1
+            total_time += time.time() - starting_time
+            if self.verbose > 0:
+                if self.epoch % self.verbose == 0:
+                    separator = "|"
+                    msg_epoch = f"| {self.epoch:<5} | "
+                    msg_epoch += f"{-fit_metrics['train']['stopping_loss']:.5f}"
+                    msg_epoch += f' {separator:<2} '
+                    msg_epoch += f"{-fit_metrics['valid']['stopping_loss']:.5f}"
+                    msg_epoch += f' {separator:<2} '
+                    msg_epoch += f" {np.round(total_time, 1):<10}"
+                    print(msg_epoch)
+
+        if self.verbose > 0:
+            if self.patience_counter == self.patience:
+                print(f"Early stopping occured at epoch {self.epoch}")
+            print(f"Training done in {total_time:.3f} seconds.")
+            print('---------------------------------------')
+
+        self.history = {"train": {"loss": self.losses_train,
+                                  "metric": self.metrics_train,
+                                  "lr": self.learning_rates},
+                        "valid": {"loss": self.losses_valid,
+                                  "metric": self.metrics_valid}}
+        # load best models post training
+        self.load_best_model()
+
+        # compute feature importance once the best model is defined
+        self._compute_feature_importances(train_dataloader)
+
+    def save_model(self, path):
+        """
+        Saving model with two distinct files.
+        """
+        saved_params = {}
+        for key, val in self.get_params().items():
+            if isinstance(val, type):
+                # Don't save torch specific params
+                continue
+            else:
+                saved_params[key] = val
+
+        # Create folder
+        Path(path).mkdir(parents=True, exist_ok=True)
+
+        # Save models params
+        with open(Path(path).joinpath("model_params.json"), "w", encoding="utf8") as f:
+            json.dump(saved_params, f)
+
+        # Save state_dict
+        torch.save(self.network.state_dict(), Path(path).joinpath("network.pt"))
+        shutil.make_archive(path, 'zip', path)
+        shutil.rmtree(path)
+        print(f"Successfully saved model at {path}.zip")
+        return f"{path}.zip"
+
+    def load_model(self, filepath):
+
+        try:
+            with zipfile.ZipFile(filepath) as z:
+                with z.open("model_params.json") as f:
+                    loaded_params = json.load(f)
+                with z.open("network.pt") as f:
+                    try:
+                        saved_state_dict = torch.load(f)
+                    except io.UnsupportedOperation:
+                        # In Python <3.7, the returned file object is not seekable (which at least
+                        # some versions of PyTorch require) - so we'll try buffering it in to a
+                        # BytesIO instead:
+                        saved_state_dict = torch.load(io.BytesIO(f.read()))
+        except KeyError:
+            raise KeyError("Your zip file is missing at least one component")
+
+        self.__init__(**loaded_params)
+
+        self.init_network(
+            input_dim=self.input_dim,
+            output_dim=self.output_dim,
+            n_d=self.n_d,
+            n_a=self.n_a,
+            n_steps=self.n_steps,
+            gamma=self.gamma,
+            cat_idxs=self.cat_idxs,
+            cat_dims=self.cat_dims,
+            cat_emb_dim=self.cat_emb_dim,
+            n_independent=self.n_independent,
+            n_shared=self.n_shared,
+            epsilon=self.epsilon,
+            virtual_batch_size=1024,
+            momentum=self.momentum,
+            device_name=self.device_name,
+            mask_type=self.mask_type
+        )
+        self.network.load_state_dict(saved_state_dict)
+        self.network.eval()
+        return
+
+    def fit_epoch(self, train_dataloader, valid_dataloader):
+        """
+        Evaluates and updates network for one epoch.
+
+        Parameters
+        ----------
+            train_dataloader: a :class: `torch.utils.data.Dataloader`
+                DataLoader with train set
+            valid_dataloader: a :class: `torch.utils.data.Dataloader`
+                DataLoader with valid set
+        """
+        train_metrics = self.train_epoch(train_dataloader)
+        valid_metrics = self.predict_epoch(valid_dataloader)
+
+        fit_metrics = {'train': train_metrics,
+                       'valid': valid_metrics}
+
+        return fit_metrics
+
+    @abstractmethod
+    def train_epoch(self, train_loader):
+        """
+        Trains one epoch of the network in self.network
+
+        Parameters
+        ----------
+            train_loader: a :class: `torch.utils.data.Dataloader`
+                DataLoader with train set
+        """
+        raise NotImplementedError('users must define train_epoch to use this base class')
+
+    @abstractmethod
+    def train_batch(self, data, targets):
+        """
+        Trains one batch of data
+
+        Parameters
+        ----------
+            data: a :tensor: `torch.tensor`
+                Input data
+            target: a :tensor: `torch.tensor`
+                Target data
+        """
+        raise NotImplementedError('users must define train_batch to use this base class')
+
+    @abstractmethod
+    def predict_epoch(self, loader):
+        """
+        Validates one epoch of the network in self.network
+
+        Parameters
+        ----------
+            loader: a :class: `torch.utils.data.Dataloader`
+                    DataLoader with validation set
+        """
+        raise NotImplementedError('users must define predict_epoch to use this base class')
+
+    @abstractmethod
+    def predict_batch(self, data, targets):
+        """
+        Make predictions on a batch (valid)
+
+        Parameters
+        ----------
+            data: a :tensor: `torch.Tensor`
+                Input data
+            target: a :tensor: `torch.Tensor`
+                Target data
+
+        Returns
+        -------
+            batch_outs: dict
+        """
+        raise NotImplementedError('users must define predict_batch to use this base class')
+
+    def load_best_model(self):
+        if self.best_network is not None:
+            self.network = self.best_network
+
+    @abstractmethod
+    def predict(self, X):
+        """
+        Make predictions on a batch (valid)
+
+        Parameters
+        ----------
+            data: a :tensor: `torch.Tensor`
+                Input data
+            target: a :tensor: `torch.Tensor`
+                Target data
+
+        Returns
+        -------
+            predictions: np.array
+                Predictions of the regression problem or the last class
+        """
+        raise NotImplementedError('users must define predict to use this base class')
+
+    def explain(self, X):
+        """
+        Return local explanation
+
+        Parameters
+        ----------
+            data: a :tensor: `torch.Tensor`
+                Input data
+            target: a :tensor: `torch.Tensor`
+                Target data
+
+        Returns
+        -------
+            M_explain: matrix
+                Importance per sample, per columns.
+            masks: matrix
+                Sparse matrix showing attention masks used by network.
+        """
+        self.network.eval()
+
+        dataloader = DataLoader(PredictDataset(X),
+                                batch_size=self.batch_size, shuffle=False)
+
+        for batch_nb, data in enumerate(dataloader):
+            data = data.to(self.device).float()
+
+            M_explain, masks = self.network.forward_masks(data)
+            for key, value in masks.items():
+                masks[key] = csc_matrix.dot(value.cpu().detach().numpy(),
+                                            self.reducing_matrix)
+
+            if batch_nb == 0:
+                res_explain = csc_matrix.dot(M_explain.cpu().detach().numpy(),
+                                             self.reducing_matrix)
+                res_masks = masks
+            else:
+                res_explain = np.vstack([res_explain,
+                                         csc_matrix.dot(M_explain.cpu().detach().numpy(),
+                                                        self.reducing_matrix)])
+                for key, value in masks.items():
+                    res_masks[key] = np.vstack([res_masks[key], value])
+        return res_explain, res_masks
+
+    def _compute_feature_importances(self, loader):
+        self.network.eval()
+        feature_importances_ = np.zeros((self.network.post_embed_dim))
+        for data, targets in loader:
+            data = data.to(self.device).float()
+            M_explain, masks = self.network.forward_masks(data)
+            feature_importances_ += M_explain.sum(dim=0).cpu().detach().numpy()
+
+        feature_importances_ = csc_matrix.dot(feature_importances_,
+                                              self.reducing_matrix)
+        self.feature_importances_ = feature_importances_ / np.sum(feature_importances_)
